MODELINKS-38: Handler and listener improvements added, covered with appropriate tests
